pyproject.toml
[tool.poetry]
name = "cfpg"
version = "0.1.0"
description = ""
authors = ["Annekin Meyburgh <annekin001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.4.1"
requests = "^2.27.1"
pytest-cov = "^3.0.0"
boltons = "^21.0.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest-integration-mark = "^0.1.0"
pytest-lazy-fixture = "^0.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
